# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LearnFlow LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /quizzes/generate:
    post:
      operationId: generateQuiz
      tags: [quizzes]
      summary: Generate a quiz for a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerateInput"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers and get results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        theme:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        theme:
          type: string
...
    QuizGenerateInput:
      type: object
      required: [videoId, videoTitle, levelTitle]
      properties:
        videoId:
          type: integer
        videoTitle:
          type: string
        levelTitle:
          type: string
    QuizQuestion:
      type: object
      required: [id, question, options, correctIndex]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
    Quiz:
      type: object
      required: [videoId, questions]
      properties:
        videoId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizAnswerInput:
      type: object
      required: [questionId, selectedIndex]
      properties:
        questionId:
          type: integer
        selectedIndex:
          type: integer
    QuizSubmission:
      type: object
      required: [videoId, videoTitle, answers]
      properties:
        videoId:
          type: integer
        videoTitle:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswerInput"
    QuizResult:
      type: object
      required: [score, totalQuestions, passed, feedback]
      properties:
        score:
          type: integer
        totalQuestions:
          type: integer
        passed:
          type: boolean
        feedback:
          type: string
        needsImprovement:
          type: boolean
...