# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnFlow LMS API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface Quiz {
  videoId: number;
  questions: QuizQuestion[];
}