# lib/api-zod/src/generated/types/quizAnswerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LearnFlow LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizAnswerInput {
  questionId: number;
  selectedIndex: number;
}